+++ client/src/hooks/MCP/index.ts
@@ -1,1 +1,3 @@
+export * from './useMCPSelect';
+export * from './useGetMCPTools';
 export { useMCPServerManager } from './useMCPServerManager';
