+++ client/src/hooks/Plugins/index.ts
@@ -1,4 +1,3 @@
-export * from './useMCPSelect';
 export * from './useToolToggle';
 export { default as useAuthCodeTool } from './useAuthCodeTool';
 export { default as usePluginInstall } from './usePluginInstall';
